pyproject.toml
[project]
name = "vlm-geometry-bench"
version = "0.1.0"
description = "Benchmark for evaluating VLM ability to identify and locate geometric shapes"
requires-python = ">=3.11"
dependencies = [
    # Image processing
    "pillow>=10.0.0",

    # Configuration
    "pyyaml>=6.0",

    # CLI and display
    "rich>=13.0.0",
    "rich-click>=1.7.0",

    # API communication
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",

    # Data handling
    "pydantic>=2.0.0",

    # Progress tracking
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
vlm-geometry-bench = "vlm_geometry_bench.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vlm_geometry_bench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
